+++ src/core/forms/winner_form.py
@@ -1,12 +1,14 @@
 from django import forms
 
-from crispy_forms.helper import FormHelper
 from crispy_forms.layout import Submit
 
+from core.forms.base_crispy_form import BaseCrispyForm
 from core.models.round import Round
 
 
-class WinnerForm(forms.ModelForm):
+class WinnerForm(BaseCrispyForm, forms.ModelForm):
+    SUBMIT_BUTTON_VALUE = "Declare Winner"
+
     class Meta:
         model = Round
         fields = ["winner"]
@@ -14,22 +16,14 @@ class WinnerForm(forms.ModelForm):
     def __init__(self, *args, players=None, moderator_id: int = None, **kwargs):
         self.players = players
         self.moderator_id = moderator_id
-        self.helper = FormHelper()
-        self.helper.form_method = "post"
-        self.helper.add_input(
-            Submit(
-                name="action",
-                value="Declare Winner",
-            )
-        )
+        super().__init__(*args, **kwargs)
+
         self.helper.add_input(
             Submit(
                 name="action",
                 value="Reveal Submitter",
             )
         )
-
-        super().__init__(*args, **kwargs)
 
         player_choices = [{"pk": None, "player_name": "-----"}] + list(players)
         self.fields["winner"].choices = (
