+++ src/core/templates/signup.html
@@ -1,23 +1,9 @@
 {% extends 'base.html' %}
+{% load crispy_forms_tags %}
 
 {% block page_content %}
   <div class="container container-flex justify-content-center">
     <h2>Sign up</h2>
-    <form method="post">
-      {% csrf_token %}
-      {% for field in form %}
-        <p>
-          {{ field.label_tag }}<br />
-          {{ field }}
-          {% if field.help_text %}
-            <small style="color: grey">{{ field.help_text }}</small>
-          {% endif %}
-          {% for error in field.errors %}
-            <p style="color: red">{{ error }}</p>
-          {% endfor %}
-        </p>
-      {% endfor %}
-      <button type="submit">Sign up</button>
-    </form>
+    {% crispy form %}
   </div>
 {% endblock %}
