+++ src/core/templates/uploads.html
@@ -1,4 +1,6 @@
-{% extends 'base.html' %} {% block page_content %}
+{% extends 'base.html' %}
+{% load crispy_forms_tags %}
+{% block page_content %}
   <div class="container-fluid">
     <div class="jumbotron">
       <h1 class="display-4">My Uploads</h1>
@@ -9,10 +11,7 @@
         <div class="card-body">
           <h5 class="card-title">Upload</h5>
           <p class="card-text">Submit your groove</p>
-          <form class="form" method="post">
-            {% csrf_token %} {{ form.as_p }}
-            <button class="btn btn-primary btn-block" type="submit">Submit Groove</button>
-          </form>
+          {% crispy form %}
         </div>
       </div>
     </div>
